Refactor major managers with helpers
